intra api doesn't work

+++ pong/main/static/home.js
@@ -3,6 +3,32 @@
     var profileImage = document.getElementById("profileImage");
     var menu = document.getElementById("menu");
 
+    const url_code = window.location.search?.slice(6)
+
+    requestData = {
+        "code": url_code
+    }
+
+    const url = `http://10.12.17.4:8000/api/v1/login/`
+    fetch(url, {
+        method: 'POST',
+        headers:
+        {
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(requestData)
+    })
+    .then(response =>
+    {
+        if (!response.ok)
+        {
+            throw new Error('Network response was not ok');
+        }
+        return response.json();
+    })
+    
+    
+    
     profileImage.addEventListener("click", function() {
       if (menu.style.display === "block") {
         menu.style.display = "none";
@@ -49,7 +75,7 @@ document.getElementById('settingsId').addEventListener('click', function(e)
     return;
   }
 
-  const url = `http://localhost:8000/api/v1/settings/${userId}/`;
+  const url = `http://10.12.17.4:8000/api/v1/settings/${userId}/`;
   fetch(url, {
       method: 'GET',
       headers: {
@@ -67,7 +93,7 @@ document.getElementById('settingsId').addEventListener('click', function(e)
     // Assuming your Django view will return a JSON response with user data
     
     console.log(data);
-    window.location.href = `http://localhost:8000/settings/`;
+    window.location.href = `http://10.12.17.4:8000/settings/`;
   })
 //   .then(response => {
 //     if (!response.ok) {
@@ -89,6 +115,44 @@ document.getElementById('settingsId').addEventListener('click', function(e)
 
 document.getElementById('logoutId').addEventListener('click', function(e)
 {
+  requested_data = {
+    "token": localStorage.getItem('access'),
+    "refresh": localStorage.getItem('refresh')
+  }
+  const token = localStorage.getItem('access');
+  if (!token)
+  {
+    alert('No token found. Please log in.');
+    window.location.href = '/';
+    return;
+  }
+  const userId = extractUserIdFromToken(token);
+  if (!userId)
+  {
+    alert('Invalid token. Please log in again.');
+    window.location.href = '/';
+    return;
+  }
+  const url = `http://10.12.17.4:8000/api/v1/logout/${userId}/`;
+  fetch(url, {
+    method: 'POST',
+    headers: {
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + token
+    },
+    body: JSON.stringify(requested_data)
+    
+  })
+  .then(response => {
+  if (!response.ok) {
+    throw new Error('Network response was not ok');
+  }
+  return response.json();
+  })
+  .then(data => {
+    console.log(data);
+  })
+
   localStorage.clear();
   window.history.pushState({}, "", '/');
   window.location.href = '/';
@@ -112,7 +176,7 @@ document.getElementById('profileId').addEventListener('click', function(e)
     return;
   }
 
-  const url = `http://localhost:8000/api/v1/profile/${userId}/`;
+  const url = `http://10.12.17.4:8000/api/v1/profile/${userId}/`;
   fetch(url, {
       method: 'GET',
       headers: {
